# docs/swagger.yaml
basePath: /
definitions:
  handler.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/handler.ErrorDetail'
      success:
        example: false
        type: boolean
    type: object
...
  handler.SuccessResponse-array_models_Order:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      success:
        example: true
        type: boolean
    type: object
  handler.SuccessResponse-handler_deleteProductResponse:
    properties:
      data:
        $ref: '#/definitions/handler.deleteProductResponse'
      success:
        example: true
        type: boolean
    type: object
  handler.SuccessResponse-handler_listProductsResponse:
    properties:
      data:
        $ref: '#/definitions/handler.listProductsResponse'
      success:
        example: true
        type: boolean
    type: object
...
  handler.SuccessResponse-models_Order:
    properties:
      data:
        $ref: '#/definitions/models.Order'
      success:
        example: true
        type: boolean
    type: object
  handler.SuccessResponse-models_Product:
    properties:
      data:
        $ref: '#/definitions/models.Product'
      success:
        example: true
        type: boolean
    type: object
...
  handler.createProductRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
    type: object
  handler.deleteProductResponse:
    properties:
      message:
        example: Product deleted successfully
        type: string
    type: object
  handler.listProductsResponse:
    properties:
      pagination:
        $ref: '#/definitions/handler.paginationInfo'
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
    type: object
...
  handler.paginationInfo:
    properties:
      current_page:
        example: 1
        type: integer
      limit:
        example: 10
        type: integer
      total_items:
        example: 25
        type: integer
    type: object
...
  handler.updateProductRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
    type: object
...
  models.Order:
    properties:
      created_at:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      status:
        type: string
      total_amount:
        description: stored in cents/paise
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.OrderItem:
    properties:
      id:
        type: integer
      order_id:
        type: integer
      price_at_purchase:
        description: stored in cents/paise
        type: integer
      product_id:
        type: integer
      product_name:
        description: populated on retrieve via join
        type: string
      quantity:
        type: integer
    type: object
  models.Product:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        description: stored in cents/paise (e.g. 99900 = 999.00)
        type: integer
      stock:
        type: integer
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: ShopFlow modern backend engineering showcase API in Go.
  title: ShopFlow API
  version: "1.0"
paths:
...
  /api/v1/orders:
    get:
      description: Retrieve history of orders placed by the authenticated user. Authenticated.
      produces:
      - application/json
      responses:
        "200":
          description: Orders retrieved successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse-array_models_Order'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - Order
    post:
      description: Convert current active cart items into a purchase order. Authenticated.
      produces:
      - application/json
      responses:
        "201":
          description: Order placed successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse-models_Order'
        "400":
          description: Cart empty or invalid data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Insufficient stock
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Place order
      tags:
      - Order
  /api/v1/orders/{id}:
    get:
      description: Retrieve full details of a specific order by ID. Authenticated.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order details retrieved successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse-models_Order'
        "400":
          description: Invalid order ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get order details
      tags:
      - Order
  /api/v1/products:
    get:
      description: Retrieve a list of products with optional pagination. Public.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Products retrieved successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse-handler_listProductsResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List products
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Create a new product. Admin only.
      parameters:
      - description: Product details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse-models_Product'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - Product
  /api/v1/products/{id}:
    delete:
      description: Delete a product by its ID. Admin only.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse-handler_deleteProductResponse'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Update an existing product. Admin only.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product update fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse-models_Product'
        "400":
          description: Invalid input or ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - Product
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your-jwt-token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"